pyproject.toml
[tool.poetry]
name = "ai-2025"
version = "0.1.0"
description = ""
authors = ["Piotr Marecki <pmarecki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.10.6"
loguru = "^0.7.3"
openai = "^1.68.2"
python-dotenv = "^1.1.0"
huggingface-hub = "^0.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
